+++ sentry-rails/spec/sentry/rails/controller_methods_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+require "sentry/rails/controller_methods"
 
 RSpec.describe Sentry::Rails::ControllerMethods do
   include described_class