+++ sentry-rails/spec/support/test_rails_app/app.rb
@@ -103,7 +103,7 @@ def make_basic_app
 
     Sentry.init do |config|
       config.release = 'beta'
-      config.dsn = "dummy://12345:67890@sentry.localdomain:3000/sentry/42"
+      config.dsn = "http://12345:67890@sentry.localdomain:3000/sentry/42"
       # for speeding up request specs
       config.rails.report_rescued_exceptions = false
       config.transport.transport_class = Sentry::DummyTransport