+++ sentry-ruby/lib/sentry/transaction.rb
@@ -17,12 +17,8 @@ module Sentry
 
       @name = name
       @parent_sampled = parent_sampled
-      set_span_recorder
-    end
-
-    def set_span_recorder
-      @span_recorder = SpanRecorder.new(1000)
-      @span_recorder.add(self)
+      @transaction = self
+      init_span_recorder
     end
 
     def self.from_sentry_trace(sentry_trace, configuration: Sentry.configuration, **options)
@@ -33,14 +29,14 @@ module Sentry
       return if match.nil?
       trace_id, parent_span_id, sampled_flag = match[1..3]
 
-      sampled =
+      parent_sampled =
         if sampled_flag.nil?
           nil
         else
           sampled_flag != "0"
         end
 
-      new(trace_id: trace_id, parent_span_id: parent_span_id, parent_sampled: sampled, sampled: sampled, **options)
+      new(trace_id: trace_id, parent_span_id: parent_span_id, parent_sampled: parent_sampled, **options)
     end
 
     def to_hash
@@ -49,20 +45,9 @@ module Sentry
       hash
     end
 
-    def start_child(**options)
-      child_span = super
-      child_span.span_recorder = @span_recorder
-
-      if @sampled
-        @span_recorder.add(child_span)
-      end
-
-      child_span
-    end
-
     def deep_dup
       copy = super
-      copy.set_span_recorder
+      copy.init_span_recorder(@span_recorder.max_length)
 
       @span_recorder.spans.each do |span|
         # span_recorder's first span is the current span, which should not be added to the copy's spans
@@ -73,7 +58,7 @@ module Sentry
       copy
     end
 
-    def set_initial_sample_decision(sampling_context: {}, configuration: Sentry.configuration)
+    def set_initial_sample_decision(sampling_context:, configuration: Sentry.configuration)
       unless configuration.tracing_enabled?
         @sampled = false
         return
@@ -81,20 +66,19 @@ module Sentry
 
       return unless @sampled.nil?
 
-      transaction_description = generate_transaction_description
-
-      logger = configuration.logger
-      sample_rate = configuration.traces_sample_rate
       traces_sampler = configuration.traces_sampler
 
-      if traces_sampler.is_a?(Proc)
-        sampling_context = sampling_context.merge(
-          parent_sampled: @parent_sampled,
-          transaction_context: self.to_hash
-        )
+      sample_rate =
+        if traces_sampler.is_a?(Proc)
+          traces_sampler.call(sampling_context)
+        elsif !sampling_context[:parent_sampled].nil?
+          sampling_context[:parent_sampled]
+        else
+          configuration.traces_sample_rate
+        end
 
-        sample_rate = traces_sampler.call(sampling_context)
-      end
+      transaction_description = generate_transaction_description
+      logger = configuration.logger
 
       unless [true, false].include?(sample_rate) || (sample_rate.is_a?(Numeric) && sample_rate >= 0.0 && sample_rate <= 1.0)
         @sampled = false
@@ -135,6 +119,13 @@ module Sentry
       hub ||= Sentry.get_current_hub
       event = hub.current_client.event_from_transaction(self)
       hub.capture_event(event)
+    end
+
+    protected
+
+    def init_span_recorder(limit = 1000)
+      @span_recorder = SpanRecorder.new(limit)
+      @span_recorder.add(self)
     end
 
     private