+++ sentry-ruby/spec/sentry_spec.rb
@@ -64,6 +64,18 @@ RSpec.describe Sentry do
 
       new_thread.join
     end
+
+    it "stores the hub in a thread variable (instead of just fiber variable)" do
+      Sentry.set_tags(outside_fiber: true)
+
+      fiber = Fiber.new do
+        Sentry.set_tags(inside_fiber: true)
+      end
+
+      fiber.resume
+
+      expect(Sentry.get_current_scope.tags).to eq({ outside_fiber: true, inside_fiber: true })
+    end
   end
 
   describe ".configure_scope" do
@@ -196,6 +208,46 @@ RSpec.describe Sentry do
           described_class.start_transaction(transaction: transaction)
 
           expect(transaction.sampled).to eq(true)
+        end
+
+        it "provides proper sampling context to the traces_sampler" do
+          transaction = Sentry::Transaction.new(op: "foo")
+
+          context = nil
+          Sentry.configuration.traces_sampler = lambda do |sampling_context|
+            context = sampling_context
+          end
+
+          described_class.start_transaction(transaction: transaction)
+
+          expect(context[:parent_sampled]).to be_nil
+          expect(context[:transaction_context][:op]).to eq("foo")
+        end
+
+        it "passes parent_sampled to the sampling_context" do
+          transaction = Sentry::Transaction.new(parent_sampled: true)
+
+          context = nil
+          Sentry.configuration.traces_sampler = lambda do |sampling_context|
+            context = sampling_context
+          end
+
+          described_class.start_transaction(transaction: transaction)
+
+          expect(context[:parent_sampled]).to eq(true)
+        end
+      end
+
+      context "when given a custom_sampling_context" do
+        it "takes that into account" do
+          context = nil
+          Sentry.configuration.traces_sampler = lambda do |sampling_context|
+            context = sampling_context
+          end
+
+          described_class.start_transaction(custom_sampling_context: { foo: "bar" })
+
+          expect(context).to include({ foo: "bar" })
         end
       end
     end