+++ sentry-delayed_job/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## 4.3.1
+
 - Add the `report_after_job_retries` configuration option to only report an exception to Sentry if this is the last job's retry after multiple exceptions. [#1364](https://github.com/getsentry/sentry-ruby/pull/1364)
+- Return delayed job when the SDK is not initialized [#1373](https://github.com/getsentry/sentry-ruby/pull/1373)
+  - Fixes [#1334](https://github.com/getsentry/sentry-ruby/issues/1334)
 
 ## 4.3.0